requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn
pydantic-settings
joblib
pandas
scikit-learn